requirements.txt
chardet==3.0.4
pdfminer.six==20200402
pycryptodome==3.9.7
sortedcontainers==2.1.0
